+++ lib/stm32/f7/Makefile
@@ -42,7 +42,7 @@ TGT_CFLAGS	+= $(STANDARD_FLAGS)
 
 ARFLAGS		= rcs
 
-OBJS		+= adc_common_v1.o adc_common_v1_multi.o
+OBJS		+= adc_common_v1.o adc_common_v1_multi.o adc_common_f47.o
 OBJS		+= can.o
 OBJS		+= crc_common_all.o crc_v2.o
 OBJS		+= dac_common_all.o
